Server/requirements.txt
alembic==1.13.2
alt-profanity-check==1.2.1
aniso8601==9.0.1
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
certifi==2022.12.7
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
datetime2==0.9.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.5
djangorestframework==3.14.0
dnspython==2.3.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
greenlet==3.0.3
idna==3.6
importlib_resources==6.4.0
ipykernel==6.29.0
ipython==8.20.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
mysql-connector-python==8.0.33
mysqlclient==2.2.4
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.24.2
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pipenv==2023.11.15
platformdirs==4.1.0
profanity-check==1.0.3
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.3.3
PyMySQL==1.1.1
pypi==2.1
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
scikit-learn==1.2.1
scipy==1.10.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
sqlparse==0.4.4
stack-data==0.6.3
termcolor==2.2.0
terminado==0.18.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.14
virtualenv==20.25.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.3
